Change how security key works
